Product insert, update and delete Done

+++ mvc/app/code/local/Core/Model/Abstract.php
@@ -104,7 +104,9 @@ class Core_Model_Abstract
     }
     public function delete()
     {
-        $this->getResource()->delete($this);
+        if($this->getId()){
+            $this->getResource()->delete($this);
+        };
         return $this;
     }
